updates
generated blender files for the amazon proposal.

intermittent_amazon_modified.blend : this is figure 3. need to change cup color scheme.

kinova_zonopy_amazon_modified.blend(1) : this is figure 1, either 1 or not 1, not sure which one. Can play with colors. Need to update transparency.

two_position_amazon_modified.blend(1) : not sure which is which but this is figure 2. need to adjust goal to solid green, start to solid gray, and change cup color scheme.